+++ insteonrf/__init__.py
@@ -13,7 +13,7 @@ from .packet import (
     pkt_crc,
 )
 
-__version__ = "2.2.0"
+__version__ = "2.3.0"
 __all__ = [
     "Address",
     "Command",